tests: avoid two testutils
